trophallaxis GT data: force data types

+++ bb_behavior/trophallaxis/evaluate_models.py
@@ -31,7 +31,9 @@ def load_test_set(root_path=None):
     trophallaxis_annotations = all_annotations[all_annotations.comment.str.contains("trophallaxis")]
     trophallaxis_annotations = trophallaxis_annotations[~pd.isnull(trophallaxis_annotations.bee_id1)]
     gt_data = trophallaxis_annotations[["frame_id", "bee_id0", "bee_id1"]]
-
+    gt_data.frame_id = gt_data.frame_id.astype(np.uint64)
+    gt_data.bee_id0 = gt_data.bee_id0.astype(np.uint16)
+    gt_data.bee_id1 = gt_data.bee_id1.astype(np.uint16)
     return annotated_frame_ids, gt_data
 
 def apply_prefilter_to_frame(frame_id, **kwargs):
